vk: rt: produce model_headers together with tlas
Makes the same producer instance produce both.
Also moves model headers buffer from vk_rtx to ray_accel, where it
belongs.
